backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
PyMuPDF>=1.24.0
chromadb==0.4.24
google-generativeai>=0.8.0
sentence-transformers>=3.0.0
supabase>=2.4.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1